+++ moo/mooutils/mooapp-ipc.c
@@ -228,7 +228,7 @@ moo_ipc_send (GObject    *sender,
               gssize      len)
 {
     GString *header;
-    gsize id_len;
+    guint id_len;
 
     g_return_if_fail (!sender || G_IS_OBJECT (sender));
     g_return_if_fail (id != NULL);
